refactor : 재사용을 위해 geocoding 기능 분리
